fix(audit): keep detector cluster in existing surfaces
