+++ src/core/code_audit/field_patterns.rs
@@ -483,6 +483,12 @@ mod tests {
     use super::*;
 
     #[test]
+    fn test_run() {
+        let dir = tempfile::tempdir().unwrap();
+        assert!(run(dir.path()).is_empty());
+    }
+
+    #[test]
     fn extracts_rust_struct_fields() {
         let content = r#"
 pub struct Config {
